+++ src/formatters/stmt_formatter.rs
@@ -483,6 +483,8 @@ impl CodeFormatter {
     }
 
     pub fn format_stmt<'ast>(&mut self, stmt: &Stmt<'ast>) -> Stmt<'ast> {
+        crate::check_should_format!(self, stmt);
+
         fmt_stmt!(self, stmt, {
             Assignment = format_assignment,
             Do = format_do_block,